factor out common logic for write ins + manual results
